refactor: add dynamic router for workspace and refactor derectory
