+++ web/src/beta/features/AccountSetting/hooks.ts
@@ -1,0 +1,41 @@
+import { useMeFetcher } from "@reearth/services/api";
+import { useCallback } from "react";
+
+type UpdatePasswordType = {
+  password: string;
+  passwordConfirmation: string;
+};
+
+export default () => {
+  const { useMeQuery, useUpdatePassword, useDeleteUser } = useMeFetcher();
+  const { me: data } = useMeQuery();
+
+  const passwordPolicy = window.REEARTH_CONFIG?.passwordPolicy;
+
+  const handleUpdateUserPassword = useCallback(
+    async ({ password, passwordConfirmation }: UpdatePasswordType) => {
+      try {
+        await useUpdatePassword({ password, passwordConfirmation });
+      } catch (error) {
+        console.error("Failed to update password:", error);
+      }
+    },
+    [useUpdatePassword]
+  );
+
+  const handleDeleteUser = useCallback(async () => {
+    try {
+      const userId = data.id;
+      if (userId) await useDeleteUser({ userId });
+    } catch (error) {
+      console.error("Failed to delete user:", error);
+    }
+  }, [data.id, useDeleteUser]);
+
+  return {
+    meData: data,
+    passwordPolicy,
+    handleUpdateUserPassword,
+    handleDeleteUser
+  };
+};
